Split out co-adding code into a separate file
