+++ reproject/mosaicking/__init__.py
@@ -1,1 +1,2 @@
+from .coadd import *  # noqa
 from .wcs_helpers import *  # noqa
